fix: count porcelain rows with the parser the tests actually cover
The daemon's countPorcelain re-scanned git status output with a laxer filter
than parsePorcelainRows: length < 3 instead of < 4, and no check that column 3
is the separator. Since the daemon collector is what feeds the sidebar's +N -M
chips, the twelve porcelain edge-case tests in test/lib/git-parsers.test.ts
were covering an implementation no longer in that path.

The parser moves to kobe-daemon/daemon/git-porcelain and countPorcelain is
built on it; kobe re-exports so its callers and tests are unchanged. Two
inputs the old counter billed as additions — a status pair with no path, and
a line with no separator in column 3 — now count as nothing.
