+++ packages/kobe-daemon/package.json
@@ -39,6 +39,7 @@
     "./daemon/deferred-prompts-store": "./src/daemon/deferred-prompts-store.ts",
     "./daemon/engine-events-log": "./src/daemon/engine-events-log.ts",
     "./daemon/event-bus": "./src/daemon/event-bus.ts",
+    "./daemon/git-porcelain": "./src/daemon/git-porcelain.ts",
     "./daemon/file-watch-trigger": "./src/daemon/file-watch-trigger.ts",
     "./daemon/issues-store": "./src/daemon/issues-store.ts",
     "./daemon/keybindings-watcher": "./src/daemon/keybindings-watcher.ts",
